requirements.txt
adal==1.2.4
azure-common==1.1.25
azure-mgmt-compute==4.6.2
azure-mgmt-network==2.7.0
azure-mgmt-resource==2.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
darksky_weather==1.9.0
haikunator==2.1.0
idna==2.10
lightgbm==2.3.1
isodate==0.6.0
jupyter==1.0.0
msrest==0.6.17
msrestazure==0.6.4
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.5
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlalchemy==1.3.18
urllib3==1.25.9
